Upgrade to rails 2.3.4 from 2.2.2
Changes simpledb to use amazon_sdb rather than aws-sdb

Reworks how the Counter store it's count, so that the count is stored in multiple values of a single attribute, rather than multiple attributes.

Adds SimpledbResource, which will be a common super-class for all models that rely on SimpleDB. Removes dependency on ActiveResource and aws-sdb-proxy.

git-svn-id: https://tapjoy.unfuddle.com/svn/tapjoy_tapjoyads/trunk@76 6df48309-bd3c-4361-8df8-ec25136b6021