Allow users to banned or be opted out from only select offer types.
Added the functionality to ban a user completely. For now this behaves exactly as an opted out customer.
Allow the user to be selectively opted out. We should be able to specify a list of offers (like EmailOffer, GenericOffer etc) for which the user is opted out. They should still see the other offers though.