Move around the checkout of deploy
This fixes the case where new nodes are picking up the current
version from the master branch instead of the deploy branch.
It's fairly simple in that moving the deploy checkout before the
version check allows it to pick up the version, since we already
always check out the deploy branch later anyway.