Re-added legacy code and created separate ET email job.
After speaking with @amdtech, we decided to break the deployment process into two stages.  First, we'll deploy the ExactTarget code and corresponding queue job.  (*Note*: the ET job is separate from the original email job.  This will allow us to manually switch back to using SendGrid without having to revert the new code.)  After we verify that the ExactTarget code is working properly, we'll clean out the legacy code again.