Uniformly handle indentifiers
Uniformly handle indentifiers, allow advertising_id, udid, and
mac_address to all act as identifiers, and have a Device's key mapped to
a tapjoy_device_id.

In the API for looking up an object from tapjoy.com Added check for Device class of an object, and do the full lookup on Device or look in DeviceIdentifier.