Began working toward adding protocol handler to action offers. Added js for an onclick handler to perform an action based on the protocol handler.
Put back the redirect on a fail protocol handler for an ios device. Right now it looks as though we will botch the ios and solely go with android till we can figure out what to do if an ios device does not have the app when a protocol handler exists.

Removed code for protocol handlers for ios. Now it's solely for android. On a click of the button the app will open, if on the user's device, or redirect to the android app store (complete_action_url).

Setup ppe offers to have protocol handlers work with android, but not ios.

Added protocol handler for ppe offers. Added view for user's who are on ios and do not have the uri scheme for that app (app is not installed), to say they do not have the app.

Finished added protocol handlers for ppe offers. Still need to iron out on what is to be done for the flow with ios devices when there is a protocol handler but the device does not have the app.

Cleaned up the offer_instructions controller with a before_filter for similar code being used between two actions. Added translations for views with the input button in index and the app_not_installed message statement.

Updated locales gem to leates version for translations on index and app_not_installed pages. In jquery set phUrl to empty string for much cleaner check on string length rather than nil. Passing app_id instead of name now.

Fixed locale text per request by johnberg, updated locales gem.

Added action_app_id to encrypted data param for security, removed unnecessary code in the app_not_installed view.

Fixed issues with calling deletion of data if offer type is not an action offer. Removed code that was added that needed to be removed (url generation action_app_id)

Added test plan for ppe protocol handlers for testing on staging servers for QA team

Added better css for buttons on offer_instructions for a cleaner look

Removed unnecessary css in the app_not_installed view.

Added acts_as_cacheable to action offers, removed memoize. Fixed tests for new find_in_cache in offer_instructions controller.

Added js to check on ios if chrome is being used. We don't want to show the button for chrome because there are issues with user coming back to the page after app is opened.

Added some more comments on the testing plan for ppe protocol handlers for QA.

Added more device types to check for

Added check for android to allow protocol handlers to be set for android platform apps.

Had to add else to if statment when checking for the type of device to show the button on android always.

Changed the way android calls the protocol handler, now using a try catch.

Refactored and added more functionality for when an android device is used. Also cleaned up ppe stuff for ios devices.

Had to change the way android opens the protocol handler, also changed if it's an action offer and a protocol handler is not present, to show a link to a complete_action_url on ios and android (it was just android before).

Fixed test for showing the protocol handler field for ios and android devices now. Added check for the protocol handler to validate that the URI scheme is valid.

Refactored the regex to check for a protocol handler. Added boundaries to the sting, so that it's all just one word, and it has to be non-whitespace.

Added new regex specifically for protocol handlers and handling the case for wanting to submit a blank handler.

Added to test plan about having to input a valid protocol handler by the user.

Added if statement wrapped around the javascript to prevent the javascript from being run when it's not a PPE offer, the action offer is not present, or protocol handler is not present.

Had to fix the checks for action offer, action app, and protocol handler, failing when nil protocol handler.

Added check to regex to prevent special characters being inputted at the begininng of the protocol handler string

Memoized the call from offer to get the associated action offer's app. Fixed tests for the changes.