Automatically cache all records as they are loaded.
To enable this feature you must be using development mode and you must set the
AUTO_CACHE_MODELS environment variable. A value of `true` will turn automatic
caching on for *all* models. You may also pass in a comma-delimated list of
models that can be cached.

Logging for this feature is extremely verbose so you can see exactly what is
or isn't getting cached, and why.