Started working on currency sales, which will allow a publisher to add a multiplier to the currency earned by a user for a time frame, and also allow for a message to make users aware that a sale is going on.
Added check on multiplier to make sure it's included in the dropdown list, cleaned up the views. Memoized all the calls from currency to currency sales from sdk.

Fixed issues with validations, and added method on controller to check for changes to end_time or start_time on an update.

Added lightbox to get_offers webpage which displays when an active currency sale is happending, or the user sets a custom message on their currency dashboard. Also added message_enabled boolean on currency to know when a custom message is going to be used. Reduced the multipliers that can be used for a currency sale.

Added jquery cookie plugin to set cookies to a currency_sale id which expire after four hours.

Added message and message enabled to currency sale as well. A currency can now have it's own message at any time to a user on the offerwall. A currency sale can either occur w/o a message/lightbox or w/ a message/lightbox, and the message can either be defaulted or set custom.